finished cleaning up, remade pipfile
