--- requirements.txt
@@ -1,34 +1,0 @@
-#
-# These requirements were autogenerated by pipenv
-# To regenerate from the project's Pipfile, run:
-#
-#    pipenv lock --requirements
-#
-
--i https://pypi.org/simple
-beautifulsoup4==4.10.0; python_version >= '3.1'
-bs4==0.0.1
-certifi==2021.5.30
-charset-normalizer==2.0.4; python_version >= '3'
-colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-configparser==5.0.2; python_version >= '3.6'
-crayons==0.4.0
-greenlet==1.1.1; python_version >= '3' and platform_machine in 'x86_64 X86_64 aarch64 AARCH64 ppc64le PPC64LE amd64 AMD64 win32 WIN32'
-idna==3.2; python_version >= '3'
-imdbpy==2021.4.18
-lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
-numpy==1.21.2; python_version < '3.11' and python_version >= '3.7'
-pandas==1.3.3
-pypiwin32==223
-python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-pytz==2021.1
-pywin32==301
-requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
-selenium==3.141.0
-six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-soupsieve==2.2.1; python_version >= '3.6'
-sqlalchemy==1.4.23; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
-tabulate==0.8.9
-urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
-webdriver-manager==3.4.2
-win10toast-persist==0.10.1
